refactor: fix some package names
does not actually affect anything
