+++ internal/cli/url/url.go
@@ -1,4 +1,4 @@
-package open
+package url
 
 import (
 	"fmt"
